# api/openapi/v1/launches.yaml
openapi: 3.0.0
...
info:
  version: 1.0.0
  title: launch-api
  description: TODO
  contact:
    name: Egor Danko
    email: qwertyone@gmail.com
servers:
  - url: 'http://{hostname}:3000/v1'
    variables:
      hostname:
        default: localhost
...
paths:
  /kinds:
    get:
      summary: List Kinds
      operationId: listKinds
      responses:
        '200':
          $ref: '#/components/responses/ListKindsResponse'
      description: Retrieves a list of kinds.
      tags:
        - kind
      parameters:
        - schema:
            type: string
          in: query
          name: cursor
        - schema:
            type: integer
            default: 20
          in: query
          name: limit
        - schema:
            type: string
            enum:
              - published
              - draft
            default: published
          in: query
          name: status
    post:
      summary: Create New Kind
      operationId: createKind
      responses:
        '204':
          description: Kind Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Kind'
              examples: {}
        '400':
          description: Missing Required Information
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '409':
          description: Kind with this name already exists
      description: Create a new kind.
      requestBody:
        $ref: '#/components/requestBodies/CreateKindRequest'
      tags:
        - kind
  '/kinds/{id}':
    parameters:
      - schema:
          type: string
          format: uuid
        name: id
        in: path
        required: true
        description: Id of an existing kind
    get:
      summary: Get Kind by ID
      tags:
        - kind
      responses:
        '200':
          $ref: '#/components/responses/GetKindResponse'
        '404':
          description: Kind Not Found
      operationId: getKind
      description: Retrieve the information of the kind with the matching kind ID.
    delete:
      summary: Delete Kind by ID
      operationId: deleteKind
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Kind Not Found
      description: Deletes existing kind.
      tags:
        - kind
  '/kinds/{id}/changeName':
    parameters:
      - schema:
          type: string
          format: uuid
        name: id
        in: path
        required: true
        description: Id of an existing kind
    post:
      summary: Change Kind Name
      operationId: changeKindName
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      requestBody:
        $ref: '#/components/requestBodies/ChangeKindName'
      tags:
        - kind
      description: Change existing kind name.
  '/kinds/{id}/changeDescription':
    parameters:
      - schema:
          type: string
          format: uuid
        name: id
        in: path
        required: true
        description: Id of an existing kind
    post:
      summary: Change Kind Description
      operationId: changeKindDescription
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      requestBody:
        $ref: '#/components/requestBodies/ChangeKindDescription'
      tags:
        - kind
      description: Change existing kind description.
  '/kinds/{id}/makePublished':
    parameters:
      - schema:
          type: string
          format: uuid
        name: id
        in: path
        required: true
        description: Id of an existing kind
    post:
      summary: Make Kind Published
      operationId: makeKindPublished
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      tags:
        - kind
      description: Change status of an existing kind to published.
  '/kinds/{id}/makeDraft':
    parameters:
      - schema:
          type: string
          format: uuid
        name: id
        in: path
        required: true
        description: Id of an existing kind
    post:
      summary: Make Kind Draft
      operationId: makeKindDraft
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      tags:
        - kind
      description: Change status of an existing kind to draft.
  /orders:
    get:
      summary: List Orders
      operationId: listOrders
      responses:
        '200':
          $ref: '#/components/responses/ListOrdersResponse'
      description: Retrieves a list of orders.
      tags:
        - order
      parameters:
        - schema:
            type: string
          in: query
          name: cursor
        - schema:
            type: integer
            default: 20
          in: query
          name: limit
        - schema:
            type: string
            enum:
              - published
              - draft
            default: published
          in: query
          name: status
    post:
      summary: Create New Order
      operationId: createOrder
      responses:
        '204':
          description: Order Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
              examples: {}
        '400':
          description: Missing Required Information
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '409':
          description: Order with this name already exists
      description: Create a new order.
      requestBody:
        $ref: '#/components/requestBodies/CreateOrderRequest'
      tags:
        - order
  '/orders/{id}':
    parameters:
      - schema:
          type: string
          format: uuid
        name: id
        in: path
        required: true
        description: Id of an existing order
    get:
      summary: Get Order by ID
      tags:
        - order
      responses:
        '200':
          $ref: '#/components/responses/GetOrderResponse'
        '404':
          description: Order Not Found
      operationId: getOrder
      description: Retrieve the information of the order with the matching order ID.
    delete:
      summary: Delete Order by ID
      operationId: deleteOrder
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Order Not Found
      description: Deletes existing order.
      tags:
        - order
  '/orders/{id}/changeName':
    parameters:
      - schema:
          type: string
          format: uuid
        name: id
        in: path
        required: true
        description: Id of an existing order
    post:
      summary: Change Order Name
      operationId: changeOrderName
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      requestBody:
        $ref: '#/components/requestBodies/ChangeOrderName'
      tags:
        - order
      description: Change existing order name.
  '/orders/{id}/changeDescription':
    parameters:
      - schema:
          type: string
          format: uuid
        name: id
        in: path
        required: true
        description: Id of an existing order
    post:
      summary: Change Order Description
      operationId: changeOrderDescription
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      requestBody:
        $ref: '#/components/requestBodies/ChangeOrderDescription'
      tags:
        - order
      description: Change existing order description.
  '/orders/{id}/makePublished':
    parameters:
      - schema:
          type: string
          format: uuid
        name: id
        in: path
        required: true
        description: Id of an existing order
    post:
      summary: Make Order Published
      operationId: makeOrderPublished
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      tags:
        - order
      description: Change status of an existing order to published.
  '/orders/{id}/makeDraft':
    parameters:
      - schema:
          type: string
          format: uuid
        name: id
        in: path
        required: true
        description: Id of an existing order
    post:
      summary: Make Order Draft
      operationId: makeOrderDraft
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      tags:
        - order
      description: Change status of an existing order to draft.
  /launches:
    get:
      summary: List Launches
      operationId: listLaunches
      responses:
        '200':
          $ref: '#/components/responses/ListLaunchesResponse'
      description: Retrieves a list of launches.
      tags:
        - launch
      parameters:
        - schema:
            type: string
          in: query
          name: cursor
        - schema:
            type: integer
            default: 20
          in: query
          name: limit
        - schema:
            type: string
            enum:
              - published
              - draft
            default: published
          in: query
          name: status
    post:
      summary: Create New Launch
      operationId: createLaunch
      responses:
        '204':
          description: Launch Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Launch'
              examples: {}
        '400':
          description: Missing Required Information
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '409':
          description: Launch with this name already exists
      description: Create a new launch.
      requestBody:
        $ref: '#/components/requestBodies/CreateLaunchRequest'
      tags:
        - launch
  '/launches/{id}':
    parameters:
      - schema:
          type: string
          format: uuid
        name: id
        in: path
        required: true
        description: Id of an existing launch
    get:
      summary: Get Launch by ID
      tags:
        - launch
      responses:
        '200':
          $ref: '#/components/responses/GetLaunchResponse'
        '404':
          description: Launch Not Found
      operationId: getLaunch
      description: Retrieve the information of the launch with the matching launch ID.
    delete:
      summary: Delete Launch by ID
      operationId: deleteLaunch
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Launch Not Found
      description: Deletes existing launch.
      tags:
        - launch
  '/launches/{id}/changeName':
    parameters:
      - schema:
          type: string
          format: uuid
        name: id
        in: path
        required: true
        description: Id of an existing launch
    post:
      summary: Change Launch Name
      operationId: changeLaunchName
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      requestBody:
        $ref: '#/components/requestBodies/ChangeLaunchName'
      tags:
        - launch
      description: Change existing launch name.
  '/launches/{id}/changeDescription':
    parameters:
      - schema:
          type: string
          format: uuid
        name: id
        in: path
        required: true
        description: Id of an existing launch
    post:
      summary: Change Launch Description
      operationId: changeLaunchDescription
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      requestBody:
        $ref: '#/components/requestBodies/ChangeLaunchDescription'
      tags:
        - launch
      description: Change existing launch description.
...
components:
  schemas:
    Kind:
      title: Kind
      type: object
      description: Kind model.
      properties:
        id:
          type: string
          description: Unique identifier for the given kind.
          format: uuid
        name:
          type: string
          description: Unique name for the given kind.
        description:
          type: string
          description: Description for the given kind.
        status:
          type: string
          enum:
            - published
            - draft
      required:
        - id
        - name
        - status
...
  responses:
    GetKindResponse:
      description: Get Kind Response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Kind'
          examples:
            Got Draft Kind:
              value:
                id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                name: kindName
                description: some description
                status: draft
    ListKindsResponse:
      description: Paginated Kinds Response
      content:
        application/json:
          schema:
            type: object
            properties:
              kinds:
                type: array
                items:
                  $ref: '#/components/schemas/Kind'
              nextCursor:
                type: string
            required:
              - kinds
              - nextCursor
...